Authentication and TestCase created

+++ src/main/java/com/example/getmesocialservice/resource/AlbumResource.java
@@ -1,11 +1,15 @@
 package com.example.getmesocialservice.resource;
 
 import com.example.getmesocialservice.model.Album;
+import com.example.getmesocialservice.model.FireBaseUser;
 import com.example.getmesocialservice.service.AlbumService;
+import com.example.getmesocialservice.service.FireBaseService;
+import com.google.firebase.auth.FirebaseAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -17,9 +21,16 @@ public class AlbumResource {
     @Autowired
     AlbumService albumService;
 
+    @Autowired
+    FireBaseService fireBaseService;
+
     @PostMapping
-    public Album saveAlbum(@RequestBody @Valid Album album){
-        return albumService.saveAlbum(album);
+    public Album saveAlbum(@RequestBody @Valid Album album, @RequestHeader("idToken") String idToken) throws IOException, FirebaseAuthException {
+        FireBaseUser fireBaseUser = fireBaseService.authenticate(idToken);
+        if (fireBaseUser != null) {
+            return albumService.saveAlbum(album);
+        }
+        return null;
     }
 
     @GetMapping
@@ -28,18 +39,23 @@ public class AlbumResource {
     }
 
     @GetMapping("/findById")
-    public Optional<Album> getAlbumById(@RequestParam(name = "albumId") String  albumId){
+    public Album getAlbumById(@RequestParam(name = "albumId") String  albumId){
         return albumService.getAlbumById(albumId);
     }
     @PutMapping
-    public Album updateAlbum(@RequestBody  @Valid Album album){
-        return albumService.updateAlbum(album);
-
+    public Album updateAlbum(@RequestBody  @Valid Album album, @RequestHeader("idToken") String idToken) throws IOException, FirebaseAuthException {
+        FireBaseUser fireBaseUser = fireBaseService.authenticate(idToken);
+        if(fireBaseUser!=null) {
+            return albumService.updateAlbum(album);
+        }
+        return null;
     }
     @DeleteMapping
-    public void deleteAlbum(String albumId){
-         albumService.deleteAlbum(albumId);
-
+    public void deleteAlbum(String albumId, @RequestHeader("idToken") String idToken ) throws IOException, FirebaseAuthException {
+        FireBaseUser fireBaseUser = fireBaseService.authenticate(idToken);
+        if(fireBaseUser!=null) {
+            albumService.deleteAlbum(albumId);
+        }
     }
 
 }
