+++ src/main/java/com/example/getmesocialservice/resource/CommentResource.java
@@ -1,11 +1,15 @@
 package com.example.getmesocialservice.resource;
 
 import com.example.getmesocialservice.model.Comment;
+import com.example.getmesocialservice.model.FireBaseUser;
 import com.example.getmesocialservice.service.CommentService;
+import com.example.getmesocialservice.service.FireBaseService;
+import com.google.firebase.auth.FirebaseAuthException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -17,10 +21,16 @@ public class CommentResource {
     @Autowired
     private CommentService commentService;
 
-    @PostMapping
-    public Comment saveComment(@RequestBody @Valid Comment comment){
-        return commentService.saveComment(comment);
+    @Autowired
+    FireBaseService fireBaseService;
 
+    @PostMapping
+    public Comment saveComment(@RequestBody @Valid Comment comment, @RequestHeader("idToken") String idToken) throws IOException, FirebaseAuthException {
+        FireBaseUser fireBaseUser = fireBaseService.authenticate(idToken);
+        if(fireBaseUser!=null) {
+            return commentService.saveComment(comment);
+        }
+        return null;
     }
     @GetMapping
     public List<Comment> getAllComments(){
@@ -33,17 +43,19 @@ public class CommentResource {
     }
 
     @PutMapping
-    public Comment updateComment(@RequestBody @Valid Comment comment) {
-        return commentService.updateComment(comment);
+    public Comment updateComment(@RequestBody @Valid Comment comment , @RequestHeader("idToken") String idToken) throws IOException, FirebaseAuthException {
+            FireBaseUser fireBaseUser = fireBaseService.authenticate(idToken);
+            if(fireBaseUser!=null) {
+                return commentService.updateComment(comment);
+            }
+            return null;
     }
 
     @DeleteMapping
-    public void deleteComment(String commentId){
-        commentService.deleteComment(commentId);
-    }
-
-
-
-
-
+    public void deleteComment(String commentId , @RequestHeader("idToken") String idToken) throws IOException, FirebaseAuthException {
+            FireBaseUser fireBaseUser = fireBaseService.authenticate(idToken);
+            if (fireBaseUser != null) {
+                commentService.deleteComment(commentId);
+            }
+        }
 }
